+++ projects/ngx-hateoas-client/src/lib/model/resource/base-resource.spec.ts
@@ -1,14 +1,14 @@
 import { BaseResource } from './base-resource';
-import { async, TestBed } from '@angular/core/testing';
+import { TestBed, waitForAsync } from '@angular/core/testing';
 import { of } from 'rxjs';
 import { ResourceHttpService } from '../../service/internal/resource-http.service';
 import { DependencyInjector } from '../../util/dependency-injector';
 import { PagedResourceCollectionHttpService } from '../../service/internal/paged-resource-collection-http.service';
 import { PagedResourceCollection } from './paged-resource-collection';
 import { ResourceCollection } from './resource-collection';
-import { HttpParams } from '@angular/common/http';
 import { ResourceCollectionHttpService } from '../../service/internal/resource-collection-http.service';
 import { LibConfig } from '../../config/lib-config';
+import { RequestParam } from '../declarations';
 
 class TestProductResource extends BaseResource {
   // tslint:disable-next-line:variable-name
@@ -60,7 +60,7 @@ describe('BaseResource GET_RELATION', () => {
   let baseResource: BaseResource;
   let resourceHttpServiceSpy: any;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     resourceHttpServiceSpy = {
       get: jasmine.createSpy('get')
     };
@@ -149,13 +149,26 @@ describe('BaseResource GET_RELATION', () => {
     });
   });
 
+  it('should undefine params/sort options for TEMPLATED link', () => {
+    resourceHttpServiceSpy.get.and.returnValue(of(new TestOrderResource()));
+
+    baseResource.getRelation('paymentType', {
+      params: {
+        projection: 'paymentProjection'
+      }
+    }).subscribe(() => {
+      const options = resourceHttpServiceSpy.get.calls.argsFor(0)[1];
+      expect(options).toEqual({...options, params: undefined, sort: undefined});
+    });
+  });
+
 });
 
 describe('BaseResource GET_RELATED_COLLECTION', () => {
   let baseResource: BaseResource;
   let resourceCollectionHttpServiceSpy: any;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     resourceCollectionHttpServiceSpy = {
       get: jasmine.createSpy('get')
     };
@@ -240,13 +253,26 @@ describe('BaseResource GET_RELATED_COLLECTION', () => {
     });
   });
 
+  it('should undefine params/sort options for TEMPLATED link', () => {
+    resourceCollectionHttpServiceSpy.get.and.returnValue(of(new TestOrderResource()));
+
+    baseResource.getRelatedCollection('paymentType', {
+      params: {
+        projection: 'paymentProjection'
+      }
+    }).subscribe(() => {
+      const options = resourceCollectionHttpServiceSpy.get.calls.argsFor(0)[1];
+      expect(options).toEqual({...options, params: undefined, sort: undefined});
+    });
+  });
+
 });
 
 describe('BaseResource GET_RELATED_PAGE', () => {
   let baseResource: BaseResource;
   let pagedResourceCollectionHttpServiceSpy: any;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     pagedResourceCollectionHttpServiceSpy = {
       get: jasmine.createSpy('get')
     };
@@ -338,7 +364,7 @@ describe('BaseResource GET_RELATED_PAGE', () => {
     });
   });
 
-  it('should pass only useCache options for TEMPLATED link', () => {
+  it('should undefine params/pageParams/sort options for TEMPLATED link', () => {
     pagedResourceCollectionHttpServiceSpy.get.and.returnValue(of(new PagedResourceCollection(new ResourceCollection())));
 
     baseResource.getRelatedPage('product', {
@@ -352,7 +378,7 @@ describe('BaseResource GET_RELATED_PAGE', () => {
       }
     }).subscribe(() => {
       const options = pagedResourceCollectionHttpServiceSpy.get.calls.argsFor(0)[1];
-      expect(options).toEqual({useCache: undefined});
+      expect(options).toEqual({...options, params: undefined, pageParams: undefined, sort: undefined});
     });
   });
 
@@ -362,7 +388,7 @@ describe('BaseResource POST_RELATION', () => {
   let baseResource: BaseResource;
   let resourceHttpServiceSpy: any;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     resourceHttpServiceSpy = {
       post: jasmine.createSpy('post')
     };
@@ -411,14 +437,25 @@ describe('BaseResource POST_RELATION', () => {
     });
   });
 
+  it('should undefine params options for TEMPLATED link', () => {
+    resourceHttpServiceSpy.post.and.returnValue(of(new TestOrderResource()));
+
+    baseResource.postRelation('updateStatusTemplated', {body: {}}, {
+      params: {statusId: 1}
+    }).subscribe(() => {
+      const options = resourceHttpServiceSpy.post.calls.argsFor(0)[2];
+      expect(options).toEqual({...options, params: undefined});
+    });
+  });
+
   it('should pass http request params when url IS NOT templated', () => {
     resourceHttpServiceSpy.post.and.returnValue(of(new TestOrderResource()));
 
     baseResource.postRelation('updateStatus', {body: {}}, {
       params: {statusId: 1}
     }).subscribe(() => {
-      const httpParams = resourceHttpServiceSpy.post.calls.argsFor(0)[2].params;
-      expect(httpParams.has('statusId')).toBeTrue();
+      const requestParams = resourceHttpServiceSpy.post.calls.argsFor(0)[2].params as RequestParam;
+      expect(requestParams.statusId).toBe(1);
     });
   });
 
@@ -467,7 +504,7 @@ describe('BaseResource POST_RELATION', () => {
 
       const options = resourceHttpServiceSpy.post.calls.argsFor(0)[2];
       expect(options).toBeDefined();
-      expect((options.params as HttpParams).keys().length).toBe(0);
+      expect(options.params).toBeUndefined();
     });
   });
 
@@ -482,7 +519,7 @@ describe('BaseResource POST_RELATION', () => {
 
       const options = resourceHttpServiceSpy.post.calls.argsFor(0)[2];
       expect(options).toBeDefined();
-      expect((options.params as HttpParams).keys().length).toBe(0);
+      expect(options.params).toBeUndefined();
     });
   });
 
@@ -492,7 +529,7 @@ describe('BaseResource PATCH_RELATION', () => {
   let baseResource: BaseResource;
   let resourceHttpServiceSpy: any;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     resourceHttpServiceSpy = {
       patch: jasmine.createSpy('patch')
     };
@@ -541,14 +578,25 @@ describe('BaseResource PATCH_RELATION', () => {
     });
   });
 
+  it('should undefine params options for TEMPLATED link', () => {
+    resourceHttpServiceSpy.patch.and.returnValue(of(new TestOrderResource()));
+
+    baseResource.patchRelation('updateStatusTemplated', {body: {}}, {
+      params: {statusId: 1}
+    }).subscribe(() => {
+      const options = resourceHttpServiceSpy.patch.calls.argsFor(0)[2];
+      expect(options).toEqual({...options, params: undefined});
+    });
+  });
+
   it('should pass http request params when url IS NOT templated', () => {
     resourceHttpServiceSpy.patch.and.returnValue(of(new TestOrderResource()));
 
     baseResource.patchRelation('updateStatus', {body: {}}, {
       params: {statusId: 1}
     }).subscribe(() => {
-      const httpParams = resourceHttpServiceSpy.patch.calls.argsFor(0)[2].params;
-      expect(httpParams.has('statusId')).toBeTrue();
+      const httpParams = resourceHttpServiceSpy.patch.calls.argsFor(0)[2].params as RequestParam;
+      expect(httpParams.statusId).toBe(1);
     });
   });
 
@@ -597,7 +645,7 @@ describe('BaseResource PATCH_RELATION', () => {
 
       const options = resourceHttpServiceSpy.patch.calls.argsFor(0)[2];
       expect(options).toBeDefined();
-      expect((options.params as HttpParams).keys().length).toBe(0);
+      expect(options.params).toBeUndefined();
     });
   });
 
@@ -612,7 +660,7 @@ describe('BaseResource PATCH_RELATION', () => {
 
       const options = resourceHttpServiceSpy.patch.calls.argsFor(0)[2];
       expect(options).toBeDefined();
-      expect((options.params as HttpParams).keys().length).toBe(0);
+      expect(options.params).toBeUndefined();
     });
   });
 
@@ -622,7 +670,7 @@ describe('BaseResource PUT_RELATION', () => {
   let baseResource: BaseResource;
   let resourceHttpServiceSpy: any;
 
-  beforeEach(async(() => {
+  beforeEach(waitForAsync(() => {
     resourceHttpServiceSpy = {
       put: jasmine.createSpy('put')
     };
@@ -665,14 +713,25 @@ describe('BaseResource PUT_RELATION', () => {
     });
   });
 
+  it('should undefine params options for TEMPLATED link', () => {
+    resourceHttpServiceSpy.put.and.returnValue(of(new TestOrderResource()));
+
+    baseResource.putRelation('updateStatusTemplated', {body: {}}, {
+      params: {statusId: 1}
+    }).subscribe(() => {
+      const options = resourceHttpServiceSpy.put.calls.argsFor(0)[2];
+      expect(options).toEqual({...options, params: undefined});
+    });
+  });
+
   it('should pass http request params when url IS NOT templated', () => {
     resourceHttpServiceSpy.put.and.returnValue(of(new TestOrderResource()));
 
     baseResource.putRelation('updateStatus', {body: {}}, {
       params: {statusId: 1}
     }).subscribe(() => {
-      const httpParams = resourceHttpServiceSpy.put.calls.argsFor(0)[2].params;
-      expect(httpParams.has('statusId')).toBeTrue();
+      const httpParams = resourceHttpServiceSpy.put.calls.argsFor(0)[2].params as RequestParam;
+      expect(httpParams.statusId).toBe(1);
     });
   });
 
@@ -710,7 +769,7 @@ describe('BaseResource PUT_RELATION', () => {
 
       const options = resourceHttpServiceSpy.put.calls.argsFor(0)[2];
       expect(options).toBeDefined();
-      expect((options.params as HttpParams).keys().length).toBe(0);
+      expect(options.params).toBeUndefined();
     });
   });
 
@@ -725,7 +784,7 @@ describe('BaseResource PUT_RELATION', () => {
 
       const options = resourceHttpServiceSpy.put.calls.argsFor(0)[2];
       expect(options).toBeDefined();
-      expect((options.params as HttpParams).keys().length).toBe(0);
+      expect(options.params).toBeUndefined();
     });
   });